HSEARCH-3920 Remove the now useless class StubMappingIndexManager
